LP862286: Fix exception on trying to start PulseAudio playback on ARM.
2011-10-03  Xerxes Ranby  <xerxes@zafena.se>
	    David Henningsson <david.henningsson@canonical.com>
	    Matthias Klose <doko@ubuntu.com>

	LP862286
	* pulseaudio/src/native/org_classpath_icedtea_pulseaudio_EventLoop.c
	  (context_change_callback):
	  Fix exception on trying to start PulseAudio playback on ARM.

+++ pulseaudio/src/native/org_classpath_icedtea_pulseaudio_EventLoop.c
@@ -108,7 +108,7 @@ static void context_change_callback(pa_context* context, void* userdata) {
     assert(cls);
     jmethodID mid = (*env)->GetMethodID(env, cls, "update", "(J)V");
     assert(mid);
-    (*env)->CallVoidMethod(env, obj, mid, pa_context_get_state(context));
+    (*env)->CallVoidMethod(env, obj, mid, (jlong) pa_context_get_state(context));
     return;
 
 }
